src/main/java/ua/com/juja/garazd/sqlcmd/controller/command/DropDatabase.java
package ua.com.juja.garazd.sqlcmd.controller.command;

import ua.com.juja.garazd.sqlcmd.model.DatabaseManager;
import ua.com.juja.garazd.sqlcmd.view.View;

public class DropDatabase implements Command {
    private DatabaseManager manager;
    private View view;

    public DropDatabase(DatabaseManager manager, View view) {
        this.manager = manager;
        this.view = view;
    }

    @Override
    public boolean canProcess(String command) {
        return command.startsWith("dropDatabase|");
    }

    @Override
    public void process(String command) {
        String[] data = command.split("\\|");
        if (data.length != 2) {
            throw new IllegalArgumentException("Command format is 'dropDatabase|databaseName', and you input: " + command);
        }

        manager.dropDatabase(data[1]);
        view.write("Base '" + data[1] + "' dropped.");
    }
}